Move URL edit box from feed info to feed settings
