+++ app/src/main/java/de/danoeh/antennapod/ui/screen/feed/preferences/FeedSettingsPreferenceFragment.java
@@ -1,6 +1,8 @@
 package de.danoeh.antennapod.ui.screen.feed.preferences;
 
 import android.Manifest;
+import android.content.ActivityNotFoundException;
+import android.content.Context;
 import android.content.Intent;
 import android.content.pm.PackageManager;
 import android.net.Uri;
@@ -13,6 +15,9 @@ import android.view.ViewGroup;
 import android.widget.Toast;
 import androidx.activity.result.ActivityResultLauncher;
 import androidx.activity.result.contract.ActivityResultContracts;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.documentfile.provider.DocumentFile;
 import androidx.core.content.ContextCompat;
 import androidx.preference.ListPreference;
 import androidx.preference.Preference;
@@ -33,9 +38,11 @@ import de.danoeh.antennapod.model.feed.VolumeAdaptionSetting;
 import de.danoeh.antennapod.net.download.serviceinterface.FeedUpdateManager;
 import de.danoeh.antennapod.storage.database.DBReader;
 import de.danoeh.antennapod.storage.database.DBWriter;
+import de.danoeh.antennapod.storage.database.FeedDatabaseWriter;
 import de.danoeh.antennapod.storage.preferences.UserPreferences;
 import de.danoeh.antennapod.ui.preferences.screen.synchronization.AuthenticationDialog;
 import de.danoeh.antennapod.ui.screen.feed.RenameFeedDialog;
+import io.reactivex.rxjava3.core.Completable;
 import io.reactivex.rxjava3.core.Maybe;
 import io.reactivex.rxjava3.core.MaybeOnSubscribe;
 import io.reactivex.rxjava3.android.schedulers.AndroidSchedulers;
@@ -62,6 +69,8 @@ public class FeedSettingsPreferenceFragment extends PreferenceFragmentCompat {
     private static final String PREF_NOTIFICATION = "episodeNotification";
     private static final String PREF_RENAME = "rename";
     private static final String PREF_TAGS = "tags";
+    private static final String PREF_EDIT_FEED_URL = "editFeedUrl";
+    private static final String PREF_RECONNECT_LOCAL_FOLDER = "reconnectLocalFolder";
 
     private Feed feed;
     private Disposable disposable;
@@ -74,6 +83,9 @@ public class FeedSettingsPreferenceFragment extends PreferenceFragmentCompat {
         fragment.setArguments(arguments);
         return fragment;
     }
+
+    private final ActivityResultLauncher<Uri> addLocalFolderLauncher =
+            registerForActivityResult(new AddLocalFolder(), this::addLocalFolderResult);
 
     boolean notificationPermissionDenied = false;
     private final ActivityResultLauncher<String> enableNotificationsRequestPermissionLauncher =
@@ -130,10 +142,12 @@ public class FeedSettingsPreferenceFragment extends PreferenceFragmentCompat {
                     updateAutoDownloadEnabledSummary();
                     updateNewEpisodesActionSummary();
 
+                    findPreference(PREF_RECONNECT_LOCAL_FOLDER).setVisible(feed.isLocalFeed());
                     if (feed.isLocalFeed()) {
                         findPreference(PREF_AUTHENTICATION).setVisible(false);
                         findPreference(PREF_AUTODOWNLOAD).setVisible(false);
                         findPreference(PREF_EPISODE_FILTER).setVisible(false);
+                        findPreference(PREF_EDIT_FEED_URL).setVisible(false);
                     }
 
                     findPreference(PREF_SCREEN).setVisible(true);
@@ -269,6 +283,29 @@ public class FeedSettingsPreferenceFragment extends PreferenceFragmentCompat {
             new RenameFeedDialog(getActivity(), feed).show();
             return true;
         });
+        findPreference(PREF_EDIT_FEED_URL).setOnPreferenceClickListener(preference -> {
+            new EditUrlSettingsDialog(getActivity(), feed) {
+                @Override
+                protected void setUrl(String url) {
+                    feed.setDownloadUrl(url);
+                }
+            }.show();
+            return true;
+        });
+        findPreference(PREF_RECONNECT_LOCAL_FOLDER).setOnPreferenceClickListener(preference -> {
+            MaterialAlertDialogBuilder alert = new MaterialAlertDialogBuilder(getContext());
+            alert.setMessage(R.string.reconnect_local_folder_warning);
+            alert.setPositiveButton(android.R.string.ok, (dialog, which) -> {
+                try {
+                    addLocalFolderLauncher.launch(null);
+                } catch (ActivityNotFoundException e) {
+                    Log.e(TAG, "No activity found. Should never happen...");
+                }
+            });
+            alert.setNegativeButton(android.R.string.cancel, null);
+            alert.show();
+            return true;
+        });
     }
 
     private void updateAutoDeleteSummary() {
@@ -327,6 +364,42 @@ public class FeedSettingsPreferenceFragment extends PreferenceFragmentCompat {
         };
         autoDownloadPreference.setSummary(summary);
         autoDownloadPreference.setValue("" + feedPreferences.getAutoDownload().code);
+    }
+
+    private void addLocalFolderResult(final Uri uri) {
+        if (uri == null) {
+            return;
+        }
+        if (feed == null) {
+            return;
+        }
+        Completable.fromAction(() -> {
+            getActivity().getContentResolver()
+                    .takePersistableUriPermission(uri, Intent.FLAG_GRANT_READ_URI_PERMISSION
+                            | Intent.FLAG_GRANT_WRITE_URI_PERMISSION);
+            DocumentFile documentFile = DocumentFile.fromTreeUri(getContext(), uri);
+            if (documentFile == null) {
+                throw new IllegalArgumentException("Unable to retrieve document tree");
+            }
+            feed.setDownloadUrl(Feed.PREFIX_LOCAL_FOLDER + uri.toString());
+            FeedDatabaseWriter.updateFeed(getContext(), feed, false);
+        })
+                .subscribeOn(Schedulers.computation())
+                .observeOn(AndroidSchedulers.mainThread())
+                .subscribe(
+                        () -> EventBus.getDefault().post(new MessageEvent(getString(android.R.string.ok))),
+                        error -> EventBus.getDefault().post(new MessageEvent(error.getLocalizedMessage())));
+    }
+
+    private static class AddLocalFolder extends ActivityResultContracts.OpenDocumentTree {
+        @NonNull
+        @Override
+        public Intent createIntent(@NonNull final Context context, @Nullable final Uri input) {
+            return super.createIntent(context, input)
+                    .addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION
+                            | Intent.FLAG_GRANT_WRITE_URI_PERMISSION
+                            | Intent.FLAG_GRANT_PERSISTABLE_URI_PERMISSION);
+        }
     }
 
     private boolean showPlaybackSpeedDialog(Preference preference) {
