+++ app/src/main/res/menu/feedinfo.xml
@@ -13,12 +13,4 @@
         android:title="@string/share_label"
         android:icon="@drawable/ic_share"
         android:visible="true" />
-    <item
-        android:id="@+id/reconnect_local_folder"
-        custom:showAsAction="collapseActionView"
-        android:title="@string/reconnect_local_folder"
-        android:visible="false" />
-    <item
-        android:id="@+id/edit_feed_url_item"
-        android:title="@string/edit_url_menu" />
 </menu>
